pyproject.toml
[project]
name = "image-processing-service"
version = "1.0.0"
description = "High-performance image processing microservice with Clean Architecture"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "Pillow>=11.0,<12",
    "python-multipart>=0.0.20",
    "alembic>=1.14,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.28,<1",
    "aiosqlite>=0.20,<1",
    "ruff>=0.8,<1",
    "mypy>=1.13,<2",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
